+++ src/main/webapp/listusuarios.jsp
@@ -24,7 +24,7 @@
                    <th scope="col">Nombre</th>
                     <th scope="col">Apellidos</th>
                     <th scope="col">Eliminar</th>
-
+                    <th scope="col">Modificar</th>
                    </tr>
                     </thead>
                     <tbody>
@@ -32,13 +32,14 @@
 <%
             for (Usuarios usuario: usuarios) {
             out.println("<tr>");
-            out.println("<th><td>"+usuario.getNombre()+"</th><td>"+usuario.getApellidos()+"</th><td><a href=delusuario?idusuario="+usuario.getIdusuario()+">Eliminar</a></th>");
+            out.println("<th><td>"+usuario.getNombre()+"</th><td>"+usuario.getApellidos()+"</th><td><a class=btn-warning href=delusuario?idusuario="+usuario.getIdusuario()+">Eliminar</a></th>");
             out.println("<tr>");
             }
 %>
                     </tbody>
                </table>
              <a class="btn btn-primary" href="menu.jsp" role="button">Menu inicio</a>
+             <a class="btn btn-primary" href="registrarse.jsp" role="button">Añadir nuevo</a>
             </body>
             </html>
