+++ src/main/webapp/menu.jsp
@@ -31,6 +31,7 @@
                     <a class="btn btn-warning" href="listusuarios.jsp" role="button">Listado Clientes</a>
                     <a class="btn btn-warning" href="listrutas.jsp" role="button">Listado Rutas</a>
 
+
                 </div>
                 <%
                 }
